Divide layers and markers styles editing GUI into a separate components
