+++ addon/components/layers-styles/simple.js
@@ -41,143 +41,6 @@ export default Ember.Component.extend({
   _markersStylesRenderer: Ember.inject.service('markers-styles-renderer'),
 
   /**
-    Available marlers-styles.
-    Initializes in component's 'init' method.
-
-    @property _availableMarkerStyles
-    @type String[]
-    @default null
-    @private
-  */
-  _availableMarkerStyles: null,
-
-  /**
-    Available markers-styles captions.
-
-    @property _availableMarkerStylesCaptions
-    @type String[]
-    @readonly
-  */
-  _availableMarkerStylesCaptions: Ember.computed('_availableMarkerStyles', 'i18n.locale', function() {
-    let availableMarkerStyles = this.get('_availableMarkerStyles');
-
-    let markerStylesCaptions = Ember.A();
-    if (Ember.isArray(availableMarkerStyles) && availableMarkerStyles.length > 0) {
-      let i18n = this.get('i18n');
-      markerStylesCaptions.pushObjects(availableMarkerStyles.map((markerStyle) => {
-        return i18n.t(`markers-styles.${markerStyle}.caption`);
-      }));
-    }
-
-    return markerStylesCaptions;
-  }),
-
-  /**
-    Selected markers-style caption.
-
-    @property _selectedMarkerStyleCaption
-    @type String
-    @default null
-    @private
-  */
-  _selectedMarkerStyleCaption: null,
-
-  /**
-    Observes changes in '_availableMarkerStylesCaptions' or 'styleSettings.style.marker.type' and computes '_selectedMarkerStyleCaption'.
-
-    @method _availableLayerStylesCaptionsOrSelectedLayerStyleDidChange
-    @private
-  */
-  _availableLayerStylesCaptionsOrSelectedLayerStyleDidChange: Ember.observer(
-    '_availableMarkerStylesCaptions.[]',
-    'styleSettings.style.marker.type',
-    function() {
-      let availableMarkerStyles = this.get('_availableMarkerStyles');
-      let selectedMarkerStyle = this.get('styleSettings.style.marker.type');
-      let selectedMarkerStyleIndex = availableMarkerStyles.findIndex((markerStyle) => {
-        return selectedMarkerStyle === markerStyle;
-      });
-
-      if (selectedMarkerStyleIndex >= 0) {
-        let markerStylesCaptions = this.get('_availableMarkerStylesCaptions');
-        this.set('_selectedMarkerStyleCaption', markerStylesCaptions.objectAt(selectedMarkerStyleIndex));
-      }
-    }
-  ),
-
-  /**
-    Observes changes in '_selectedMarkerStyleCaption' property and changes style type in related marker style settings hash.
-
-    @method _selectedMarkerStyleCaptionDidChange
-    @private
-  */
-  _selectedMarkerStyleCaptionDidChange: Ember.observer('_selectedMarkerStyleCaption', function () {
-    Ember.run.once(this, '_setSelectedMarkerStyle');
-  }),
-
-  /**
-    Sets selected markers-style by its i18n-ed caption.
-
-    @method _setSelectedMarkerStyle
-    @private
-  */
-  _setSelectedMarkerStyle() {
-    let availableMarkerStyles = this.get('_availableMarkerStyles');
-    let availableMarkerStylesCaptions = this.get('_availableMarkerStylesCaptions');
-    let selectedMarkerStyleCaption = this.get('_selectedMarkerStyleCaption');
-
-    if (!Ember.isArray(availableMarkerStyles) || !Ember.isArray(availableMarkerStylesCaptions) || Ember.isBlank(selectedMarkerStyleCaption)) {
-      return null;
-    }
-
-    let selectedMarkerStyleIndex = availableMarkerStylesCaptions.findIndex((markerStylesCaption) => {
-      return markerStylesCaption.toString() === selectedMarkerStyleCaption.toString();
-    });
-
-    let selectedMarkerStyle = selectedMarkerStyleIndex > -1 ?
-      availableMarkerStyles.objectAt(selectedMarkerStyleIndex) :
-      null;
-
-    this.set('styleSettings.style.marker.type', selectedMarkerStyle);
-  },
-
-  /**
-    Observes changes in 'styleSettings.style.marker.type' and computes default style setings.
-
-    @method _selectedMarkerStyleDidChange
-    @private
-  */
-  _selectedMarkerStyleDidChange: Ember.observer('styleSettings.style.marker.type', function() {
-    Ember.run.once(this, '_setSelectedMarkerStyleDefaultSettings');
-  }),
-
-  /**
-    Previosly selected layer style.
-
-    @property _previouslySelectedMarkerStyle
-    @type string
-    @private
-  */
-  _previouslySelectedMarkerStyle: null,
-
-  /**
-    Sets default style settings with respect to 'styleSettings.style.marker.type'.
-
-    @method _setSelectedMarkerStyleDefaultSettings
-    @private
-  */
-  _setSelectedMarkerStyleDefaultSettings() {
-    let previouslySelectedMarkerStyle = this.get('_previouslySelectedMarkerStyle');
-    let selectedMarkerStyle = this.get('styleSettings.style.marker.type');
-    if (Ember.isBlank(selectedMarkerStyle) || previouslySelectedMarkerStyle === selectedMarkerStyle) {
-      return;
-    }
-
-    this.set('styleSettings.style.marker.style', this.get('_markersStylesRenderer').getDefaultStyleSettings(selectedMarkerStyle));
-    this.set('_previouslySelectedMarkerStyle', selectedMarkerStyle);
-  },
-
-  /**
     Reference to component's template.
   */
   layout,
@@ -199,18 +62,6 @@ export default Ember.Component.extend({
     @default null
   */
   styleSettings: null,
-
-  /**
-    Initializes component.
-  */
-  init() {
-    this._super(...arguments);
-
-    // Initialize available markers-styles and related properties.
-    this.set('_previouslySelectedMarkerStyle', this.get('styleSettings.style.marker.type'));
-    this.set('_availableMarkerStyles', this.get('_markersStylesRenderer').getAvailableMarkerStylesTypes());
-    this._availableLayerStylesCaptionsOrSelectedLayerStyleDidChange();
-  },
 
   /**
     Initializes DOM-related component's properties.
