+++ addon/components/layers-styles/simple/path-editor.js
@@ -1,0 +1,28 @@
+/**
+  @module ember-flexberry-gis
+*/
+
+import Ember from 'ember';
+import layout from '../../../templates/components/layers-styles/simple/path-editor';
+
+/**
+  Component containing GUI for 'simple' layers-style 'path' settings.
+
+  @class SimpleLayersStylePathEditorComponent
+  @extends <a href="http://emberjs.com/api/classes/Ember.Component.html">Ember.Component</a>
+*/
+export default Ember.Component.extend({
+  /**
+    Reference to component's template.
+  */
+  layout,
+
+  /**
+    Hash containing style settings.
+
+    @property styleSettings
+    @type Object
+    @default null
+  */
+  styleSettings: null
+});
