+++ addon/locales/en/components/layers-styles-editor.js
@@ -1,0 +1,8 @@
+export default {
+  'layer-style-dropdown': {
+    'caption': 'Symbol type'
+  },
+  'layer-style-component-not-implemented': {
+    'message': 'Component \'{{componentName}}\' isn\'t implemented'
+  }
+};
