+++ addon/locales/en/components/layers-styles/empty.js
@@ -1,0 +1,3 @@
+export default {
+  'layer-will-be-invisible-message': 'Layer will be invisible on map'
+};
