+++ addon/locales/en/components/layers-styles/simple.js
@@ -1,0 +1,11 @@
+import PathEditor from './simple/path-editor';
+
+export default {
+  'path-tab': {
+    'caption': 'Linear & polygonal objects style'
+  },
+  'marker-tab': {
+    'caption': 'Point objects style'
+  },
+  'path-editor': PathEditor
+};
