+++ addon/locales/en/components/layers-styles/simple/path-editor.js
@@ -1,0 +1,2 @@
+export default {
+};
