+++ addon/locales/en/components/markers-styles-editor.js
@@ -1,0 +1,8 @@
+export default {
+  'markers-styles-dropdown': {
+    'caption': 'Marker type'
+  },
+  'marker-style-component-not-implemented': {
+    'message': 'Component \'{{componentName}}\' isn\'t implemented'
+  }
+};
