+++ addon/locales/en/components/markers-styles/default.js
@@ -1,0 +1,3 @@
+export default {
+  'default-marker-will-be-used-message': 'Default marker will be used'
+};
