+++ addon/locales/en/components/markers-styles/image.js
@@ -1,0 +1,2 @@
+export default {
+};
