+++ addon/locales/en/layers-styles/empty.js
@@ -1,4 +1,3 @@
 export default {
-  'caption': 'No symbol',
-  'layer-will-be-invisible-message': 'Layer will be invisible on map'
+  'caption': 'No symbol'
 };
