+++ addon/locales/en/layers-styles/simple.js
@@ -1,15 +1,3 @@
 export default {
-  'caption': 'Simple symbol',
-  'path-tab': {
-    'caption': 'Linear & polygonal objects style'
-  },
-  'marker-tab': {
-    'caption': 'Point objects style',
-    'markers-styles-dropdown': {
-      'caption': 'Marker type'
-    },
-    'marker-style-component-not-implemented': {
-      'message': 'Component \'{{componentName}}\' isn\'t implemented'
-    }
-  }
+  'caption': 'Simple symbol'
 };
