+++ addon/locales/en/markers-styles/default.js
@@ -1,4 +1,3 @@
 export default {
-  'caption': 'Default marker',
-  'default-marker-will-be-used-message': 'Default marker will be used'
+  'caption': 'Default marker'
 };
