+++ addon/locales/ru/components/layers-styles-editor.js
@@ -1,0 +1,8 @@
+export default {
+  'layer-style-dropdown': {
+    'caption': 'Тип знака'
+  },
+  'layer-style-component-not-implemented': {
+    'message': 'Компонент \'{{componentName}}\' не реализован'
+  }
+};
