+++ addon/locales/ru/components/layers-styles.js
@@ -1,0 +1,7 @@
+import EmptyLayerStyle from './layers-styles/empty';
+import SimpleLayerStyle from './layers-styles/simple';
+
+export default {
+  'empty': EmptyLayerStyle,
+  'simple': SimpleLayerStyle
+};
