+++ addon/locales/ru/components/layers-styles/empty.js
@@ -1,0 +1,3 @@
+export default {
+  'layer-will-be-invisible-message': 'Слой буден невидим на карте'
+};
