+++ addon/locales/ru/components/layers-styles/simple.js
@@ -1,0 +1,11 @@
+import PathEditor from './simple/path-editor';
+
+export default {
+  'path-tab': {
+    'caption': 'Стиль линейных и полигональных объектов'
+  },
+  'marker-tab': {
+    'caption': 'Стиль точечных объектов'
+  },
+  'path-editor': PathEditor
+};
