+++ addon/locales/ru/components/layers-styles/simple/path-editor.js
@@ -1,0 +1,2 @@
+export default {
+};
