+++ addon/locales/ru/components/markers-styles-editor.js
@@ -1,0 +1,8 @@
+export default {
+  'markers-styles-dropdown': {
+    'caption': 'Тип маркера'
+  },
+  'marker-style-component-not-implemented': {
+    'message': 'Компонент \'{{componentName}}\' не реализован'
+  }
+};
