+++ addon/locales/ru/components/markers-styles.js
@@ -1,0 +1,7 @@
+import DefaultMarkerStyle from './markers-styles/default';
+import ImageMarkerStyle from './markers-styles/image';
+
+export default {
+  'default': DefaultMarkerStyle,
+  'image': ImageMarkerStyle
+};
