+++ addon/locales/ru/components/markers-styles/default.js
@@ -1,0 +1,3 @@
+export default {
+  'default-marker-will-be-used-message': 'Будет использован маркер по умолчанию'
+};
