+++ addon/locales/ru/components/markers-styles/image.js
@@ -1,0 +1,2 @@
+export default {
+};
