+++ addon/locales/ru/layers-styles/empty.js
@@ -1,4 +1,3 @@
 export default {
-  'caption': 'Без знака',
-  'layer-will-be-invisible-message': 'Слой будет невидим на карте'
+  'caption': 'Без знака'
 };
