+++ addon/locales/ru/layers-styles/simple.js
@@ -1,15 +1,3 @@
 export default {
-  'caption': 'Простой знак',
-  'path-tab': {
-    'caption': 'Стиль для линейных и полигональных объектов'
-  },
-  'marker-tab': {
-    'caption': 'Стиль для точечных объектов',
-    'markers-styles-dropdown': {
-      'caption': 'Тип маркера'
-    },
-    'marker-style-component-not-implemented': {
-      'message': 'Компонент \'{{componentName}}\' не реализован'
-    }
-  }
+  'caption': 'Простой знак'
 };
