+++ addon/locales/ru/markers-styles/default.js
@@ -1,4 +1,3 @@
 export default {
-  'caption': 'Маркер по умолчанию',
-  'default-marker-will-be-used-message': 'Будет использован маркер по умолчанию'
+  'caption': 'Маркер по умолчанию'
 };
