+++ addon/templates/components/layers-styles-editor.hbs
@@ -1,0 +1,21 @@
+<div class="field">
+  <label>
+    {{t "components.layers-styles-editor.layer-style-dropdown.caption"}}
+  </label>
+  {{flexberry-dropdown
+    class="fluid"
+    items=_availableLayerStylesCaptions
+    value=_selectedLayerStyleCaption
+  }}
+</div>
+{{#if (known-for-type "component" (concat "layers-styles/" styleSettings.type))}}
+  {{component (concat "layers-styles/" styleSettings.type) styleSettings=styleSettings}}
+{{else}}
+  {{ui-message
+    type="error"
+    message=(t
+      "components.layers-styles-editor.layer-style-component-not-implemented.message"
+      componentName=(concat "layers-styles/" styleSettings.type)
+    )
+  }}
+{{/if}}
