+++ addon/templates/components/layers-styles/empty.hbs
@@ -1,3 +1,3 @@
 {{ui-message
-  message=(t "layers-styles.empty.layer-will-be-invisible-message")
+  message=(t "components.layers-styles.empty.layer-will-be-invisible-message")
 }}
