+++ addon/templates/components/layers-styles/simple/path-editor.hbs
@@ -1,0 +1,3 @@
+{{flexberry-jsonarea
+  value=styleSettings
+}}
