+++ addon/templates/components/markers-styles/default.hbs
@@ -1,3 +1,3 @@
 {{ui-message
-  message=(t "markers-styles.default.default-marker-will-be-used-message")
+  message=(t "components.markers-styles.default.default-marker-will-be-used-message")
 }}
