+++ app/components/layers-styles-editor.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/components/layers-styles-editor';
