+++ app/components/layers-styles/simple/path-editor.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/components/layers-styles/simple/path-editor';
