+++ app/components/markers-styles-editor.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/components/markers-styles-editor';
